API/requirements.txt
fastapi
openai
tiktoken
uvicorn
google-generativeai
python-dotenv
pydantic
ollama